Simplify project structure: merge some small cxx files, reduce directory nesting
