+++ src/sdp_solve/SDP/read_block_data/Block_Data_Parse_Result.hxx
