+++ src/sdp_solve/SDP/read_block_data/SDP_Block_Data.cxx
@@ -1,7 +1,7 @@
 #include "SDP_Block_Data.hxx"
 
 #include "Json_Block_Data_Parser.hxx"
-#include "sdp_solve/SDP/SDP/set_bases_blocks.hxx"
+#include "sdp_solve/SDP/set_bases_blocks.hxx"
 #include "sdpb_util/Vector_State.hxx"
 #include "sdpb_util/assert.hxx"
 #include "sdpb_util/boost_serialization.hxx"
